mbm: refactor - alloc plan to top level package
