+++ pkg/agent/qrm-plugins/mb/allocator/plan_allocator_test.go
@@ -17,11 +17,11 @@ limitations under the License.
 package allocator
 
 import (
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/plan"
 	"testing"
 
 	"github.com/stretchr/testify/mock"
 
-	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/plan"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/qosgroup"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/resctrl"
 )
