+++ pkg/agent/qrm-plugins/mb/controller/policy/strategy/ccdtarget/logarithmic_scale_planner.go
@@ -3,8 +3,8 @@ package ccdtarget
 import (
 	"math"
 
-	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/controller/policy/plan"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/monitor/stat"
+	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/plan"
 	"github.com/kubewharf/katalyst-core/pkg/agent/qrm-plugins/mb/qosgroup"
 	"github.com/kubewharf/katalyst-core/pkg/util/syntax"
 )
